refactored variables from hooks
